+++ src/TRegx/CleanRegex/Internal/Model/Match/IndexedRawMatchOffset.php
@@ -1,0 +1,19 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Model\Match;
+
+class IndexedRawMatchOffset extends RawMatchOffset
+{
+    /** @var int */
+    private $index;
+
+    public function __construct(array $match, int $index)
+    {
+        parent::__construct($match);
+        $this->index = $index;
+    }
+
+    public function getIndex(): int
+    {
+        return $this->index;
+    }
+}