+++ src/TRegx/CleanRegex/Internal/Model/Matches/IRawMatchesGroupable.php
@@ -1,0 +1,7 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Model\Matches;
+
+interface IRawMatchesGroupable
+{
+    public function matched(): bool;
+}