+++ test/Unit/CleanRegex/Internal/Model/Match/IndexedRawMatchOffsetTest.php
@@ -1,0 +1,23 @@
+<?php
+namespace Test\CleanRegex\Internal\Model\Match;
+
+use PHPUnit\Framework\TestCase;
+use TRegx\CleanRegex\Internal\Model\Match\IndexedRawMatchOffset;
+
+class IndexedRawMatchOffsetTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function shouldGetIndex()
+    {
+        // given
+        $match = new IndexedRawMatchOffset([], 14);
+
+        // when
+        $index = $match->getIndex();
+
+        // then
+        $this->assertEquals(14, $index);
+    }
+}